+++ tests/Unit/Validate/AuxiliarFolios13/ValidateTest.php
@@ -6,22 +6,22 @@ namespace PhpCfdi\CeUtils\Tests\Unit\Validate\AuxiliarFolios13;
 
 use PhpCfdi\CeUtils\AuxiliarFoliosCreator13;
 use PhpCfdi\CeUtils\Tests\TestCase;
-use PhpCfdi\CeUtils\Tests\Traits\WithFakeFiel;
+use PhpCfdi\CeUtils\Tests\Traits\WithFakeCsd;
 
 final class ValidateTest extends TestCase
 {
-    use WithFakeFiel;
+    use WithFakeCsd;
 
     public function testValidateMinimalCreatedDocument(): void
     {
-        $fiel = $this->buildFiel();
+        $credential = $this->buildCredential();
         $creator = new AuxiliarFoliosCreator13([
-            'RFC' => $fiel->certificate()->rfc(),
+            'RFC' => $credential->certificate()->rfc(),
             'Mes' => '01',
             'Anio' => '2021',
             'TipoSolicitud' => 'AF',
         ]);
-        $creator->addSello($fiel);
+        $creator->addSello($credential);
         $asserts = $creator->validate();
         $this->assertEmpty($asserts->nones());
         $this->assertFalse($asserts->hasErrors());