+++ docs/CHANGELOG.md
@@ -12,7 +12,7 @@ In summary, [SemVer](https://semver.org/) can be viewed as ` Breaking . Feature 
 
 ## Version 0.1.1 2020-07-07
 
-- Fix XSLT locations to 1.3 specs.
+- Fix XSLT locations to 1.3 specs, see <https://www.gob.mx/sat/acciones-y-programas/contabilidad-electronica-173700>.
 - Test source string after sign documents.
 - Add exception when empty source string.
 - Use CSD instead of FIEL to sign documents.